+++ src/0pages/JoinPage.tsx
@@ -1,0 +1,7 @@
+import UserInfoForm from "../1containers/UserInfoForm";
+
+const JoinPage = () => {
+  return <UserInfoForm />;
+};
+
+export default JoinPage;
