+++ src/2components/1TextField.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from "react";
-import { InputProps } from "../types/InputProps";
-import useInput from "../hooks/useInput";
+import { InputProps } from "../3types/InputProps";
+import useInput from "../4hooks/useInput";
 
 const TextField: FunctionComponent<InputProps> = ({
   validates,
@@ -8,11 +8,14 @@ const TextField: FunctionComponent<InputProps> = ({
   label,
   placeholder,
   type,
-  emailCheck,
 }) => {
   const { value, onChange, error } = useInput({ source, validates });
+  // error = {
+  //   name: '더 써주세요',
+  //   password: '더 써주세요'
+  // }
 
-  console.log("error in Textfield", error);
+  //key가 뭐 올지 모를때는(source가 변수(변함)) .말고 []로 객체에 접근
   return (
     <div>
       <div style={{ display: "flex", gridGap: "8px" }}>
@@ -23,10 +26,14 @@ const TextField: FunctionComponent<InputProps> = ({
           name={source}
           type={type}
           placeholder={placeholder}
-          onKeyUp={emailCheck}
+          // onKeyUp={emailCheck}
         />
       </div>
-      {error && <p style={{ color: "red" }}>{error}</p>}
+      <div style={{ marginTop: "12px" }}>
+        {error?.[source] && (
+          <p style={{ color: "red", marginTop: "4px" }}>{error[source]}</p>
+        )}
+      </div>
     </div>
   );
 };
