+++ src/4hooks/useInput.tsx
@@ -1,6 +1,6 @@
-import { InputProps } from "../types/InputProps";
+import { InputProps } from "../3types/InputProps";
 import React, { useCallback, useContext } from "react";
-import { FormContext } from "../components/SimpleForm";
+import { FormContext } from "../2components/0SimpleForm";
 
 interface UseInputProps extends Pick<InputProps, "source" | "validates"> {}
 
@@ -8,22 +8,24 @@ function useInput(props: UseInputProps) {
   const { setValues, values, setError, error } = useContext(FormContext);
 
   const onChange = useCallback(
-    (v: string | number) => {
+    (value: string | number) => {
       //[min(5), max(10)]
-      let error = null;
       //func는 min or max
-      props.validates.forEach((func) => {
-        error = func(v);
-        console.log("error in UseInput", error);
-
-        if (error) {
-          setError(props.source, error);
+      props.validates.forEach((validate) => {
+        const err: string = validate(value);
+        //props.source를 누가정하냐.....
+        //[]해야되는 이유....
+        if (err) {
+          setError({
+            ...error,
+            [props.source]: err,
+          });
         }
       });
 
       setValues({
         ...values,
-        [props.source]: v,
+        [props.source]: value,
       });
     },
     [values, props.source, props.validates, setError, setValues]
@@ -32,7 +34,7 @@ function useInput(props: UseInputProps) {
   return {
     value: values[props.source],
     onChange,
-    error: error[props.source],
+    error,
   };
 }
