+++ src/5validations/index.ts
@@ -18,3 +18,6 @@ export const max = (num: number) => (v: string) => {
 // "2", ""
 // setError("2")
 // setError()
+
+//try0. else값에 '올바른 입력랎입니다'를 넣어주고 싶었지만, min, max가
+//minNum < v.length < maxNum으로 만들지 못해서 안됨
