+++ src/main.tsx
@@ -2,7 +2,7 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import App from "./App";
 import "./index.css";
-import JoinPage from "./pages/JoinPage";
+import JoinPage from "./0pages/JoinPage";
 
 const root = ReactDOM.createRoot(document.getElementById("root")!);
