--- src/pages/JoinPage.tsx
@@ -1,9 +1,0 @@
-import UserInfoForm from "../containers/UserInfoForm";
-
-const JoinPage = () => {
-    return (
-        <UserInfoForm/>
-    );
-};
-
-export default JoinPage;
